+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-[unreleased]
+[2.0.3-alpha.5]
 
 * Fix network data proceeding when using custom protocol multiple times #1847
 
 * Backward incompatible change: use pre shared key as connection protector in libp2p. Add libp2p psk to invitation link
+
+* Removed code responsible for data translation from channel address to channel id from state manager transforms and storage service
 
 [2.0.3-alpha.1]